refactor test file into directory
